Move Access Management feature to vertical-slice folders
Final pilot for the Access Management slice (4 endpoints — list all
events+groups, list user accesses with pending, request access,
list+approve/reject access requests). The leftover Controllers/
EventsController.cs is renamed to Features/AccessManagement/
AccessController.cs; the only Events-routed endpoints already moved
with the Events slice.

- IAccessService and IAccessManagementService move out of
  Domain.Services into Application.Features.AccessManagement.
- AccessService and AccessManagementService move under the same
  feature folder. AccessManagementService was previously registered
  in Infrastructure DI; its registration consolidates into
  AddAccessManagementFeature() alongside IAccessService.
- AccessRoutes extracted from ApiEndpoints.cs (Video.Access nested
  class removed); AccessRoutes lives at /api/videos/accesses/* to
  preserve existing URLs unchanged.
- Four DTOs (ApproveAccessRequest, RequestAssigmentModelRequest,
  EventsAndGroupsResponse, RequestedAccessesResponse) move to
  Contracts/Features/AccessManagement/.
- Backend consumers of IAccessService (Sharing, Comments,
  VideoService, VideoController, Events controller, VideoServiceTests,
  SharedLinkServiceTests) gain usings for the new namespace.
- ContractMappers and the Access controller pick up the new
  Contracts.Features.AccessManagement namespace.
- Mobile (Library, GetAccessPageModel, test project) gain a using
  for the new Contracts.Features.AccessManagement namespace.
