+++ src/backend/Application/Features/AccessManagement/AccessManagementModule.cs
@@ -1,0 +1,16 @@
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Application.Features.AccessManagement;
+
+public static class AccessManagementModule
+{
+    extension(IServiceCollection services)
+    {
+        public IServiceCollection AddAccessManagementFeature()
+        {
+            services.AddScoped<IAccessService, AccessService>();
+            services.AddScoped<IAccessManagementService, AccessManagementService>();
+            return services;
+        }
+    }
+}
