+++ src/backend/Application/Features/Comments/CommentService.cs
@@ -1,4 +1,5 @@
 using Application.Extensions;
+using Application.Features.AccessManagement;
 using Domain.Entities;
 using Domain.Services;
 using Microsoft.EntityFrameworkCore;
