+++ src/backend/Application/Features/Sharing/SharedLinkService.cs
@@ -1,3 +1,4 @@
+using Application.Features.AccessManagement;
 using Domain.Entities;
 using Domain.Services;
 using Microsoft.EntityFrameworkCore;
