+++ src/backend/Application/Services/VideoService.cs
@@ -1,4 +1,5 @@
-﻿using Domain;
+﻿using Application.Features.AccessManagement;
+using Domain;
 using Domain.Entities;
 using Domain.Models;
 using Domain.Services;
