+++ src/backend/TB.DanceDance.API/Controllers/VideoController.cs
@@ -1,4 +1,5 @@
-﻿using Domain.Services;
+﻿using Application.Features.AccessManagement;
+using Domain.Services;
 using Infrastructure.Identity.IdentityResources;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
