+++ src/backend/TB.DanceDance.API/Features/Events/EventsController.cs
@@ -1,3 +1,4 @@
+using Application.Features.AccessManagement;
 using Application.Features.Events;
 using Domain.Services;
 using Infrastructure.Identity.IdentityResources;
