+++ src/tests/TB.DanceDance.Tests/Application/VideoServiceTests.cs
@@ -1,4 +1,5 @@
-﻿using Application.Services;
+﻿using Application.Features.AccessManagement;
+using Application.Services;
 using Domain;
 using Domain.Entities;
 using Domain.Models;
