+++ src/tests/TB.DanceDance.Tests/Features/Sharing/SharedLinkServiceTests.cs
@@ -1,3 +1,4 @@
+using Application.Features.AccessManagement;
 using Application.Features.Sharing;
 using Application.Services;
 using Domain.Entities;
